+++ .env.sample
@@ -3,4 +3,7 @@ LLM_API_KEY =
 RESEND_API_KEY = 
 DOMAIN =
 NEXT_AUTH_SECRET = 
+REDIS_HOST=
+REDIS_PORT=
+REDIS_PASSWORD=
