+++ package.json
@@ -25,6 +25,7 @@
     "clsx": "^2.1.1",
     "dompurify": "^3.1.4",
     "input-otp": "^1.2.4",
+    "ioredis": "^5.4.1",
     "lucide-react": "^0.378.0",
     "marked": "^12.0.2",
     "mongoose": "^8.4.0",
